// src/imports/interfaces/arr-metadata.interface.ts
export interface ArrMetadata {
  stationName: string;

  stationCode: string;

  installationYear?: number;

  operationYear?: number;

  operator?: string;

  province?: string;

  regency?: string;

  district?: string;

  village?: string;

  latitude?: string;

  longitude?: string;

  readingMethod?: string;
}